Changes: - renamed packages of token flow APIs - reverted VariableKeySetUriTokenDecoder back to NOT inheriting from JwtDecoder, as this would interfere with auto-configuration of xsuaaJwtDecoder.
