+++ spring-xsuaa/src/test/java/com/sap/cloud/security/xsuaa/token/flows/RestTemplateMock.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.security.xsuaa.tokenflows;
+package com.sap.cloud.security.xsuaa.token.flows;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertNotNull;
