+++ spring-xsuaa/src/test/java/com/sap/cloud/security/xsuaa/token/flows/TokenDecoderMock.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.security.xsuaa.tokenflows;
+package com.sap.cloud.security.xsuaa.token.flows;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
@@ -7,7 +7,7 @@ import java.net.URI;
 
 import org.springframework.security.oauth2.jwt.Jwt;
 
-import com.sap.cloud.security.xsuaa.tokenflows.VariableKeySetUriTokenDecoder;
+import com.sap.cloud.security.xsuaa.token.flows.VariableKeySetUriTokenDecoder;
 
 public class TokenDecoderMock implements VariableKeySetUriTokenDecoder {
