+++ spring-xsuaa/src/test/java/com/sap/cloud/security/xsuaa/token/flows/XsuaaTokenFlowsTests.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.security.xsuaa.tokenflows;
+package com.sap.cloud.security.xsuaa.token.flows;
 
 import static org.junit.Assert.*;
 
@@ -12,6 +12,11 @@ import java.util.Map;
 import org.junit.Test;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.client.RestTemplate;
+
+import com.sap.cloud.security.xsuaa.token.flows.ClientCredentialsTokenFlow;
+import com.sap.cloud.security.xsuaa.token.flows.RefreshTokenFlow;
+import com.sap.cloud.security.xsuaa.token.flows.UserTokenFlow;
+import com.sap.cloud.security.xsuaa.token.flows.XsuaaTokenFlows;
 
 public class XsuaaTokenFlowsTests {
